allow larger binary sends

+++ EtherCard.h
@@ -110,7 +110,7 @@ public:
   BufferFiller (uint8_t* buf) : start (buf), ptr (buf) {}
       
   void emit_p (PGM_P fmt, ...);
-  void emit_raw (const char* s, uint8_t n) { memcpy(ptr, s, n); ptr += n; }
+  void emit_raw (const char* s, uint16_t n) { memcpy(ptr, s, n); ptr += n; }
   
   uint8_t* buffer () const { return start; }
   uint16_t position () const { return ptr - start; }
